Add player job system and remove skillbooks
